# Streamline and improve performance for orderable sls version parsing