… sls-versions/src/main/java/com/palantir/sls/versions/MatchResult.java
… sls-versions/src/main/java/com/palantir/sls/versions/Parser.java
… sls-versions/src/main/java/com/palantir/sls/versions/RegexParser.java
… sls-versions/src/main/java/com/palantir/sls/versions/ReleaseVersionParser.java