@aldexis 2025-08-29
While this might indeed improve performance, there is no indication so far that this is needed anywhere. As such, I'll close this PR